chore: moved component to ui/cards
