+++ src/renderer/entities/wallet/ui/cards/WalletCardSm/WalletCardSm.tsx
